Document backend modular structure
