+++ backend/routes/__init__.py
@@ -1,0 +1,4 @@
+from .base import base_bp
+from .damage import damage_bp
+
+__all__ = ["base_bp", "damage_bp"]
