+++ backend/services/__init__.py
@@ -1,0 +1,3 @@
+from .damage_detection import decode_image, detect_damage
+
+__all__ = ["decode_image", "detect_damage"]
