+++ backend/services/damage_detection.py
